+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">ModularizationTest</string>
     <string name="title_dynamicfeature">Help DFM</string>
-    <string name="title_authdynamic">Authenfication</string>
+    <string name="title_auth">Authenfication</string>
     <string name="title_home">Home</string>
     <string name="home">Home</string>
     <string name="device">Device</string>
