+++ feature/auth/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    id("timHome.android.library")
+    id("timHome.android.libraryCompose")
+    id("timHome.quality.convention.plugin")
+    id("kotlin-parcelize")
+}
+android {
+    namespace = "com.example.auth"
+}
+
+dependencies {
+    implementation(project(":core"))
+    implementation(libs.firebase.analytics)
+    debugImplementation(libs.androidx.ui.tooling)
+}
