+++ feature/auth/src/main/java/com/example/auth/data/AuthorizationRepositoryImpl.kt
@@ -1,0 +1,61 @@
+package com.example.auth.data
+
+import com.example.auth.data.api.AuthApi
+import com.example.auth.data.api.request.UserLoginRequest
+import com.example.auth.domain.IAuthorizationRepository
+import com.example.auth.ui.signin.LoginViewState
+import com.example.core.data.AppDatabase
+import com.example.core.utils.Constant.CODE_200
+import com.example.core.utils.mvi.ErrorType
+import com.example.core.utils.mvi.MviError
+import javax.inject.Inject
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.onStart
+
+internal class AuthorizationRepositoryImpl
+    @Inject
+    constructor(
+        private val apiService: AuthApi,
+        private val database: AppDatabase,
+    ) : IAuthorizationRepository {
+        override fun loginByEmail(
+            oldState: LoginViewState,
+            email: String,
+            password: String,
+        ): Flow<LoginViewState> =
+            flow {
+                val body = UserLoginRequest(email, password)
+                try {
+                    val response = apiService.loginByEmail(body)
+                    if (response.code() == CODE_200) {
+                        response.body()?.user?.convertToUserEntity()
+                            .also {
+                                database.userDao().saveUserToDB(it!!)
+                            }.also {
+                                emit(oldState.copy(isLoading = false, error = null, isLoginSuccess = true))
+                            }
+                    } else {
+                        emit(
+                            oldState.copy(
+                                isLoading = false,
+                                error = MviError(type = ErrorType.TOAST, errorMessage = "loginByEmail"),
+                            ),
+                        )
+                    }
+                } catch (e: Exception) {
+                    emit(
+                        oldState.copy(
+                            isLoading = false,
+                            error =
+                                MviError(
+                                    type = ErrorType.TOAST,
+                                    errorMessage = e.message ?: "loginByEmail",
+                                ),
+                        ),
+                    )
+                }
+            }.onStart {
+                emit(oldState.copy(isLoading = true, error = null))
+            }
+    }
