+++ feature/auth/src/main/java/com/example/auth/data/api/AuthApi.kt
@@ -1,0 +1,18 @@
+package com.example.auth.data.api
+
+import com.example.auth.data.api.request.NewUserRequest
+import com.example.auth.data.api.request.UserLoginRequest
+import com.example.auth.data.api.response.LoginResponse
+import com.example.auth.data.api.response.UserResponse
+import retrofit2.Response
+import retrofit2.http.Body
+import retrofit2.http.POST
+
+internal interface AuthApi {
+    @POST("auth/login")
+    suspend fun loginByEmail(
+        @Body body: UserLoginRequest,
+    ): Response<LoginResponse>
+
+    fun newAccount(body: NewUserRequest): UserResponse
+}
