+++ feature/auth/src/main/java/com/example/auth/data/api/request/NewUserRequest.kt
@@ -1,0 +1,3 @@
+package com.example.auth.data.api.request
+
+internal data class NewUserRequest(val email: String, val userName: String, val password: String)
