+++ feature/auth/src/main/java/com/example/auth/data/api/request/UserLoginRequest.kt
@@ -1,0 +1,3 @@
+package com.example.auth.data.api.request
+
+internal data class UserLoginRequest(val email: String, val password: String)
