+++ feature/auth/src/main/java/com/example/auth/data/mapper/User.kt
@@ -1,0 +1,3 @@
+package com.example.auth.data.mapper
+
+internal data class User(val id: Int, val email: String, val userName: String, val avatar: String?)
