+++ feature/auth/src/main/java/com/example/auth/ui/signin/LoginViewState.kt
@@ -1,0 +1,26 @@
+package com.example.auth.ui.signin
+
+import android.os.Parcelable
+import com.example.core.data.models.FieldText
+import com.example.core.utils.mvi.MviError
+import com.example.core.utils.mvi.MviViewState
+import kotlinx.parcelize.Parcelize
+
+@Parcelize
+data class LoginViewState(
+    val email: FieldText = FieldText(data = null, error = null),
+    val password: FieldText = FieldText(data = null, error = null),
+    val isLoading: Boolean,
+    val isLoginBtnEnable: Boolean,
+    val isLoginSuccess: Boolean = false,
+    override val error: MviError?,
+) : MviViewState, Parcelable {
+    companion object {
+        fun initial() =
+            LoginViewState(
+                isLoading = false,
+                isLoginBtnEnable = false,
+                error = null,
+            )
+    }
+}
