+++ feature/auth/src/main/java/com/example/auth/ui/signin/SignInEvent.kt
@@ -1,0 +1,7 @@
+package com.example.auth.ui.signin
+
+internal sealed interface SignInEvent {
+    object GoToHomeScreen : SignInEvent
+
+    data class ShowErrorMessage(val error: String) : SignInEvent
+}
