+++ feature/auth/src/main/java/com/example/auth/ui/signin/di/SignInModule.kt
@@ -1,0 +1,17 @@
+package com.example.auth.ui.signin.di
+
+import androidx.lifecycle.ViewModel
+import com.example.auth.ui.signin.SignInViewModel
+import com.example.core.utils.viewmodel.ViewModelAssistedFactory
+import com.example.core.utils.viewmodel.ViewModelKey
+import dagger.Binds
+import dagger.Module
+import dagger.multibindings.IntoMap
+
+@Module
+internal interface SignInModule {
+    @Binds
+    @IntoMap
+    @ViewModelKey(SignInViewModel::class)
+    fun bindVMFactory(f: SignInViewModel.Factory): ViewModelAssistedFactory<out ViewModel>
+}
