+++ feature/auth/src/main/java/com/example/auth/ui/signup/SignUpFragment.kt
@@ -1,0 +1,3 @@
+package com.example.auth.ui.signup
+
+internal class SignUpFragment
