--- feature/authDynamic/src/main/java/com/example/authdynamic/data/api/request/NewUserRequest.kt
@@ -1,3 +1,0 @@
-package com.example.authdynamic.data.api.request
-
-internal data class NewUserRequest(val email: String, val userName: String, val password: String)
