--- feature/authDynamic/src/main/java/com/example/authdynamic/data/api/request/UserLoginRequest.kt
@@ -1,3 +1,0 @@
-package com.example.authdynamic.data.api.request
-
-internal data class UserLoginRequest(val email: String, val password: String)
