--- feature/authDynamic/src/main/java/com/example/authdynamic/data/mapper/User.kt
@@ -1,3 +1,0 @@
-package com.example.authdynamic.data.mapper
-
-internal data class User(val id: Int, val email: String, val userName: String, val avatar: String?)
