--- feature/authDynamic/src/main/java/com/example/authdynamic/ui/signin/SignInEvent.kt
@@ -1,7 +1,0 @@
-package com.example.authdynamic.ui.signin
-
-internal sealed interface SignInEvent {
-    object GoToHomeScreen : SignInEvent
-
-    data class ShowErrorMessage(val error: String) : SignInEvent
-}
