--- feature/authDynamic/src/main/java/com/example/authdynamic/ui/signup/SignUpFragment.kt
@@ -1,3 +1,0 @@
-package com.example.authdynamic.ui.signup
-
-internal class SignUpFragment
