+++ settings.gradle.kts
@@ -26,7 +26,7 @@ include(":core")
 include(":benchmark")
 include(":mock")
 include(":base")
-include(":feature:authDynamic")
+include(":feature:auth")
 include(":feature:home")
 include(":feature:device")
 include(":feature:settings")
